Added link getter to Organization model.
