+++ opengever/contact/models/organization.py
@@ -1,5 +1,6 @@
 from opengever.base.model import CONTENT_TITLE_LENGTH
 from opengever.contact.models.contact import Contact
+from opengever.contact.utils import get_contactfolder_url
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
@@ -18,3 +19,14 @@ class Organization(Contact):
     persons = relationship("OrgRole", back_populates="organization")
 
     __mapper_args__ = {'polymorphic_identity': 'organization'}
+
+    @property
+    def wrapper_id(self):
+        return 'organization-{}'.format(self.organization_id)
+
+    def get_url(self, view='view'):
+        return '{}/{}/{}'.format(
+            get_contactfolder_url(), self.wrapper_id, view)
+
+    def get_title(self, view='view'):
+        return self.name
